a) Shapefile Editor proceeded; b) background tile-loading threads will be terminated with TerminateThread call, rather then waiting for HTTP response (no delay during application shutdown);
git-svn-id: https://mapwingis.svn.codeplex.com/svn@75089 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ trunk/COM classes/Shapefile_Edit.cpp
@@ -419,7 +419,7 @@ STDMETHODIMP CShapefile::EditUpdateShape(long shapeIndex, IShape* shpNew, VARIAN
 		return S_FALSE;
 	}
 
-	if (shapeIndex < 0 || shapeIndex >= _shapeData.size())
+	if (shapeIndex < 0 || shapeIndex >= (long)_shapeData.size())
 	{
 		ErrorMessage(tkINDEX_OUT_OF_BOUNDS);
 		return S_FALSE;
@@ -440,7 +440,7 @@ void CShapefile::ReregisterShape(int shapeIndex)
 {
 	if (!_isEditingShapes) return;
 
-	if (shapeIndex < 0  || shapeIndex >= _shapeData.size())
+	if (shapeIndex < 0  || shapeIndex >= (int)_shapeData.size())
 		return;
 
 	IShape* shp = _shapeData[shapeIndex]->shape;
